% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotslapmeg.R
\name{plotslapmeg}
\alias{plotslapmeg}
\title{Plot the estimated random effects from SLaPMEG seperated for the study groups}
\usage{
plotslapmeg(obj, ...)
}
\arguments{
\item{obj}{An slapmeg object which is the output from \code{\link{slapmeg}} or
\code{\link{pairslapmeg}}. Note that the \code{fullreturn=TRUE}) must have been used.}

\item{\dots}{optional graphical parameters can be added with an \code{+} based on \code{\link{ggplot2}}
structure.}
}
\value{
returns NULL
}
\description{
This plot can provide a graphical insight into the source of effect (i.e. diffrential expression)
in relevent pathway.
}
\examples{

\dontrun{
# run slapmeg for a pathway
#' m1 <- slapmeg()
#Plot the estimated random effects
plotslapmeg(m1)
}
}
\references{
paper title goes here
}
\seealso{
\code{\link{slapmeg}}, \code{\link{multslapmeg}}, \code{\link{pairslapmeg}}
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
