% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mslapmeg.R
\name{print.mslapmeg}
\alias{print.mslapmeg}
\title{Print objects from \code{multslapmeg}}
\usage{
\method{print}{mslapmeg}(x, ...)
}
\arguments{
\item{x}{an object inheriting from class \code{multslapmeg}}

\item{\dots}{further arguments to be passed to or from other methods, which will be ignored in
this function.}
}
\value{
Returns result summary of mSLAPMEG approach
}
\description{
Provides an overview of mSLaPMEG results for each pathway through \code{multslapmeg} object.
}
\seealso{
\code{\link{multslapmeg}},\code{\link{slapmeg}}, \code{\link{pairslapmeg}}
}
\author{
Mitra Ebrahimpoor
}
\keyword{print}
