% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.slapmeg.R
\name{print.slapmeg}
\alias{print.slapmeg}
\title{Print objects from \code{slapmeg}}
\usage{
\method{print}{slapmeg}(x, ...)
}
\arguments{
\item{x}{an object inheriting from class \code{slapmeg}}

\item{\dots}{further arguments to be passed to or from other methods, which will be ignored in
this function.}
}
\value{
Returns result summaries of SLAPMEG approach
}
\description{
Provides an overview of SLaPMEG approach through \code{slapmeg}
and \code{pairslapmeg} objects.
}
\seealso{
\code{\link{slapmeg}}, \code{\link{pairslapmeg}}
}
\author{
Mitra Ebrahimpoor
}
\keyword{print}
