% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slapGT.R
\name{slapGT}
\alias{slapGT}
\title{Run step II via globaltest package}
\usage{
slapGT(EBS, data_inf, rand_nam, Ynames, grouping, Emethod)
}
\arguments{
\item{EBS}{Estimated random effects by either joint or pairwise approach}

\item{data_inf}{Dataframe with variables required for the test}

\item{rand_nam}{Names of random effects defined in model}

\item{Ynames}{Name of omics in pathway}

\item{grouping}{Name of the covariate representing the grouping by the phenotype}

\item{Emethod}{Estimation method for the random effects which is either joint or pairwise}
}
\value{
A GT object containing globaltest results
}
\description{
Globaltest is used to test the association between phenotype variable and
estimated random effects from step I.
}
\author{
Mitra Ebrahimpoor

\email{m.ebrahimpoor@lumc.nl}
}
