% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mslapmeg.R
\name{summary.mslapmeg}
\alias{summary.mslapmeg}
\title{Summary of  \code{slapmeg} objects}
\usage{
\method{summary}{mslapmeg}(object, n = 5, ...)
}
\arguments{
\item{object}{an object inheriting from classes \code{multslapmeg}}

\item{n}{an integer indicating number of pathways to be printed the default is 5}

\item{\dots}{further arguments to be passed to or from other methods, which will be ignored in this
function.}
}
\value{
'Returns the fixed effect estimates and random effects variances from the model in step 1
and the details of Globaltest along with p-value from step 2.
}
\description{
Provides a table of sorted p-values for the output of multSLaPMEG approach within \code{multslapmeg} objects.
}
\seealso{
\code{\link{multslapmeg}}, \code{\link{pairslapmeg}},
}
\author{
Mitra Ebrahimpoor
}
\keyword{summary}
