% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.slapmeg.R
\name{summary.slapmeg}
\alias{summary.slapmeg}
\title{Summary of  \code{slapmeg} objects}
\usage{
\method{summary}{slapmeg}(object, ...)
}
\arguments{
\item{object}{an object inheriting from classes \code{slapmeg}}

\item{\dots}{further arguments to be passed to or from other methods, which will be ignored in this
function.}
}
\value{
'Returns the fixed effect estimates and random effects variances from the model in step 1
and the details of Globaltest along with p-value from step 2.
}
\description{
Provides a summary of values estimated with SLaPMEG approach within \code{slapmeg}
and \code{pairslapmeg} objects.
}
\seealso{
\code{\link{slapmeg}}, \code{\link{pairslapmeg}},
}
\author{
Mitra Ebrahimpoor
}
\keyword{summary}
